requirements.txt
# Requirements to install
python-jenkins==1.7.0
pyyaml==5.4.1
